requirements.txt
pandas
sqlalchemy
yfinance
plotly
streamlit
apscheduler
pytest
python-dotenv
psycopg2-binary
